Update benchmark runner to benchmark
