--- test/perf/all.ts
@@ -1,2 +1,0 @@
-import "./propertyGet";
-import "./serdes";
